requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
altair>=5.0.0
mlb-statsapi>=1.7.2
pybaseball>=2.2.6
